refactor(scale-runners): split ec2 scale-up provider
