+++ lambdas/functions/control-plane/src/scale-runners/scale-up-config.ts
@@ -1,0 +1,8 @@
+import type { ScaleUpRunnerProviderType } from './scale-up-provider';
+
+export function getScaleUpRunnerProviderType(
+  type: string | undefined,
+  defaultType: ScaleUpRunnerProviderType,
+): ScaleUpRunnerProviderType {
+  return type ?? defaultType;
+}
